Move review form processing functions from utils.py to transfer_forms.py
